Убрана избыточная проверка. Изменен тип поля в модели Review

+++ api_yamdb/reviews/models.py
@@ -104,7 +104,7 @@ class Review(models.Model):
         related_name='reviews',
         on_delete=models.CASCADE,
     )
-    score = models.PositiveIntegerField(
+    score = models.PositiveSmallIntegerField(
         verbose_name='Оценка',
         validators=(
             MinValueValidator(1, 'Минимальная оценка - 1'),
